server/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
opencv-python
pandas
requests
